Move public classes and interfaces out from .core package
MOB-4324
