+++ widgetssdk/src/main/java/com/glia/widgets/call/CallContract.kt
@@ -4,7 +4,7 @@ import com.glia.widgets.base.BaseController
 import com.glia.widgets.base.BaseView
 import com.glia.widgets.core.dialog.model.ConfirmationDialogLinks
 import com.glia.widgets.core.dialog.model.Link
-import com.glia.widgets.core.engagement.MediaType
+import com.glia.widgets.engagement.MediaType
 import com.glia.widgets.locale.LocaleString
 
 internal interface CallContract {
