+++ widgetssdk/src/main/java/com/glia/widgets/call/domain/HandleCallPermissionsUseCase.kt
@@ -1,6 +1,6 @@
 package com.glia.widgets.call.domain
 
-import com.glia.widgets.core.engagement.MediaType
+import com.glia.widgets.engagement.MediaType
 import com.glia.widgets.core.permissions.PermissionManager
 import com.glia.widgets.engagement.domain.IsCurrentEngagementCallVisualizerUseCase
 import com.glia.widgets.permissions.PermissionsGrantedCallback
