+++ widgetssdk/src/main/java/com/glia/widgets/core/permissions/PermissionManager.kt
@@ -9,7 +9,7 @@ import android.provider.Settings
 import androidx.annotation.VisibleForTesting
 import com.glia.androidsdk.comms.MediaDirection
 import com.glia.androidsdk.comms.MediaUpgradeOffer
-import com.glia.widgets.core.engagement.MediaType
+import com.glia.widgets.engagement.MediaType
 import com.glia.widgets.helper.Logger
 import com.glia.widgets.helper.TAG
 import com.glia.widgets.permissions.Permissions
