+++ widgetssdk/src/main/java/com/glia/widgets/core/secureconversations/domain/SecureConversationTopBannerVisibilityUseCase.kt
@@ -1,7 +1,7 @@
 package com.glia.widgets.core.secureconversations.domain
 
-import com.glia.widgets.core.engagement.MediaType
-import com.glia.widgets.core.queue.Queue
+import com.glia.widgets.engagement.MediaType
+import com.glia.widgets.queue.Queue
 import com.glia.widgets.core.queue.QueueRepository
 import com.glia.widgets.core.queue.QueuesState
 import com.glia.widgets.helper.mediaTypes
