+++ widgetssdk/src/main/java/com/glia/widgets/di/GliaCore.kt
@@ -21,7 +21,7 @@ import com.glia.androidsdk.visitor.Authentication
 import com.glia.androidsdk.visitor.VisitorInfo
 import com.glia.androidsdk.visitor.VisitorInfoUpdateRequest
 import com.glia.widgets.core.authentication.AuthenticationManager
-import com.glia.widgets.core.engagement.MediaType
+import com.glia.widgets.engagement.MediaType
 import java.io.InputStream
 import java.util.Optional
 import java.util.function.Consumer
