+++ widgetssdk/src/main/java/com/glia/widgets/di/GliaCoreImpl.kt
@@ -22,8 +22,8 @@ import com.glia.androidsdk.visitor.Authentication
 import com.glia.androidsdk.visitor.VisitorInfo
 import com.glia.androidsdk.visitor.VisitorInfoUpdateRequest
 import com.glia.widgets.core.authentication.AuthenticationManager
-import com.glia.widgets.core.engagement.MediaType
-import com.glia.widgets.core.queue.toCoreType
+import com.glia.widgets.engagement.MediaType
+import com.glia.widgets.queue.toCoreType
 import java.io.InputStream
 import java.util.Optional
 import java.util.function.Consumer
