+++ widgetssdk/src/main/java/com/glia/widgets/engagement/States.kt
@@ -4,7 +4,6 @@ import com.glia.androidsdk.Engagement.ActionOnEnd
 import com.glia.androidsdk.Operator
 import com.glia.androidsdk.engagement.EngagementState
 import com.glia.androidsdk.engagement.Survey
-import com.glia.widgets.core.engagement.MediaType
 
 internal sealed interface State {
     data object NoEngagement : State
