+++ widgetssdk/src/main/java/com/glia/widgets/entrywidget/EntryWidgetController.kt
@@ -4,8 +4,8 @@ import android.app.Activity
 import androidx.annotation.VisibleForTesting
 import com.glia.widgets.chat.Intention
 import com.glia.widgets.chat.domain.IsAuthenticatedUseCase
-import com.glia.widgets.core.engagement.MediaType
-import com.glia.widgets.core.queue.Queue
+import com.glia.widgets.engagement.MediaType
+import com.glia.widgets.queue.Queue
 import com.glia.widgets.core.queue.QueueRepository
 import com.glia.widgets.core.queue.QueuesState
 import com.glia.widgets.core.secureconversations.SecureConversationsRepository
