+++ widgetssdk/src/test/java/com/glia/widgets/call/domain/HandleCallPermissionsUseCaseTest.kt
@@ -1,6 +1,6 @@
 package com.glia.widgets.call.domain
 
-import com.glia.widgets.core.engagement.MediaType
+import com.glia.widgets.engagement.MediaType
 import com.glia.widgets.core.permissions.PermissionManager
 import com.glia.widgets.engagement.domain.IsCurrentEngagementCallVisualizerUseCase
 import com.glia.widgets.permissions.Permissions
