+++ widgetssdk/src/test/java/com/glia/widgets/core/secureconversations/domain/IsMessagingAvailableUseCaseTest.kt
@@ -1,8 +1,8 @@
 package com.glia.widgets.core.secureconversations.domain
 
 import android.COMMON_EXTENSIONS_CLASS_PATH
-import com.glia.widgets.core.engagement.MediaType
-import com.glia.widgets.core.queue.Queue
+import com.glia.widgets.engagement.MediaType
+import com.glia.widgets.queue.Queue
 import com.glia.widgets.core.queue.QueueRepository
 import com.glia.widgets.core.queue.QueuesState
 import com.glia.widgets.engagement.EngagementRepository
