+++ widgetssdk/src/test/java/com/glia/widgets/entrywidget/EntryWidgetControllerTest.kt
@@ -5,8 +5,8 @@ import android.app.Activity
 import com.glia.androidsdk.Engagement
 import com.glia.androidsdk.engagement.EngagementState
 import com.glia.widgets.chat.domain.IsAuthenticatedUseCase
-import com.glia.widgets.core.engagement.MediaType
-import com.glia.widgets.core.queue.Queue
+import com.glia.widgets.engagement.MediaType
+import com.glia.widgets.queue.Queue
 import com.glia.widgets.core.queue.QueueRepository
 import com.glia.widgets.core.queue.QueuesState
 import com.glia.widgets.core.secureconversations.SecureConversationsRepository
